ASB-31860 moved test from test_operator to test_keyword
